Adopt ruff across the codebase: lint, format, and CI gate
Configure ruff 0.15.22 in pyproject.toml (E,W,F,I,UP,B,C4,SIM,ISC,PIE,PT,
RUF,NPY,PERF,PLC,PLE,PLW; line-length 100; formatter on) and bring the whole
tree into compliance. Roughly 1,700 lint findings fixed and 336 files
reformatted, then the residual manual findings resolved by hand.

The ignore list is deliberate and commented at the source: E402 and PLC0415
because prefer_pyside6() and env setup must run before Qt/backend imports and
lazy imports are load-bearing; PLW0108 because signal-connect lambdas swallow
Qt's emitted arguments by design; plus E501, ISC001, PLW2901, PLW0603, PT011,
PERF203, and RUF001-003.

Notable substantive changes surfaced by the pass:
- Fix a latent bug in the colormap toolbar: the menu could be torn down while
  the user was browsing it when the family changed mid-sync; the rebuild now
  defers until the menu closes.
- Import MontageTileOverlay from its home module (display.overlays) instead of
  a transitive re-export through display.imageview2d.
- Chain re-raised exceptions (raise ... from) in the IO and export paths.
- Declare explicit __hash__ = None on the mutable lifecycle set views, and
  narrow overly-broad exception suppressions where a specific type sufficed.

Add a fast CI `lint` job (ruff check + format --check, version-pinned) and
document the workflow and ignore rationale in AGENTS.md. Re-baseline the
frame_controller line-count guard for formatter rewrapping (no logic growth)
and make the vispy inheritance guard robust to parenthesized imports.

Full offscreen suite green: 2504 passed, 53 skipped, 1 xfailed.

+++ arrayscope/display/histogram_plot.py
@@ -12,7 +12,6 @@ from time import perf_counter
 
 import numpy as np
 
-
 MIN_HISTOGRAM_BIN_SCREEN_PX = 5
 DEFAULT_HISTOGRAM_BIN_CAP = 500
 DEFAULT_HISTOGRAM_TARGET_IMAGE_SIZE = 200
@@ -84,7 +83,7 @@ def compute_histogram_plot(request: HistogramPlotRequest) -> HistogramPlotResult
                         visible_bin_width = visible_span / max_visible_bins
                         requested_bins = max(
                             2,
-                            int(ceil(span / max(visible_bin_width, np.finfo(float).eps))),
+                            ceil(span / max(visible_bin_width, np.finfo(float).eps)),
                         )
                         cap = max(2, min(int(request.bin_cap), int(sampled.size)))
                         bins = max(2, min(requested_bins, cap))
@@ -106,14 +105,16 @@ def compute_histogram_plot(request: HistogramPlotRequest) -> HistogramPlotResult
     )
 
 
-def sample_histogram_data(data: np.ndarray, *, target_image_size: int = DEFAULT_HISTOGRAM_TARGET_IMAGE_SIZE) -> np.ndarray:
+def sample_histogram_data(
+    data: np.ndarray, *, target_image_size: int = DEFAULT_HISTOGRAM_TARGET_IMAGE_SIZE
+) -> np.ndarray:
     array = np.asarray(data)
     if np.iscomplexobj(array):
         array = np.abs(array).astype(np.float32, copy=False)
     if array.ndim < 2:
         return array.reshape(-1)
-    step0 = max(1, int(ceil(array.shape[0] / float(target_image_size))))
-    step1 = max(1, int(ceil(array.shape[1] / float(target_image_size))))
+    step0 = max(1, ceil(array.shape[0] / float(target_image_size)))
+    step1 = max(1, ceil(array.shape[1] / float(target_image_size)))
     return array[::step0, ::step1].reshape(-1)
